+++ cliar/utils.py
@@ -11,6 +11,7 @@ def set_help(help_map: Dict[str, str]) -> Callable:
     '''
     def decorator(handler: Callable) -> Callable:
         '''Decorator returning command handler with a help message map.'''
+
         handler._help_map = help_map
         return handler
 
@@ -26,6 +27,7 @@ def set_metavars(metavar_map: Dict[str, str]) -> Callable:
     '''
     def decorator(handler: Callable) -> Callable:
         '''Decorator returning command handler with a custom metavar map.'''
+
         handler._metavar_map = metavar_map
         return handler
 
@@ -44,6 +46,23 @@ def set_arg_map(arg_map: Dict[str, str]) -> Callable:
         '''Decorator returning command handler with a custom arg map.'''
 
         handler._arg_map = arg_map
+        return handler
+
+    return decorator
+
+
+def set_sharg_map(sharg_map: Dict[str, str]) -> Callable:
+    '''Override mapping from handler params to short commandline arg names.
+
+    Be default, the first character of arg names are used as short arg names.
+
+    :param arg_map: mapping from handler param names to short arg names
+    '''
+
+    def decorator(handler: Callable) -> Callable:
+        '''Decorator returning command handler with a custom shaarg map.'''
+
+        handler._sharg_map = sharg_map
         return handler
 
     return decorator