+++ tests/test_case_sensitive_args/case_sensitive_args.py
@@ -1,0 +1,11 @@
+from cliar import Cliar, set_sharg_map
+
+
+class CaseSensitiveArgs(Cliar):
+    @set_sharg_map({'url': None, 'username': 'U'})
+    def _root(self, unique='', url='', username=''):
+        pass
+
+
+if __name__ == "__main__":
+    CaseSensitiveArgs().parse()